Properly handle self-types in reflection member lookup
The type on which we lookup things needs to be the this-type of the
class, otherwise it will not include members that only appear in the
self-type, since these members can appear in the types of members of the
current class, this could lead to a MissingType exception being thrown.

Fixes #12081.