Map opaque types in arguments of inlined calls to proxies
An argument of an inlined call might have a type that refers to opaque
types in the inlined method's scope. In that case, we need to cast those
arguments to types where the opaque references are replaced by proxies,
so that the opaque aliases are visible when accessing the argument.

Fixes #12914