Support on-demand Symbol.defTree synchronization
In the current compiler, the defTree associated with a symbol becomes
out-of-sync when we transform the trees. This creates a problem for
all analysis that depend on Symbol.defTree.

An analysis that depend on tasty should perform the analysis
immediately after the picklerPhases --- as it has to handle trees
loaded from Tasty which are without any lowering.

Analysis that do not cross boundaries of compilation units
(e.g. linters) are not subject to the restriction above.