Use `StringConcatFactory` for string concatenation on JDK 9+
JEP 280, released in JDK 9, proposes a new way to compile string
concatenation using `invokedynamic` and `StringConcatFactory`.
This new approach generates less bytecode, doesn't have to incur
the overhead of `StringBuilder` allocations, and allows users to
pick swap the concatenation technique at runtime.

This changes the codegen when the target is at least Java 9 to
leverage `invokedynamic` and `StringConcatFactory`. On Java 8,
the old `StringBuilder` approach is still used.

Ported from scala#9556