Add missing case for finding default getters
There are two modes for finding default getters. They depend on whether
the receiver is defined or not. If the receiver is not defined we search
the enclosing scopes instead. This fails if the method and the call are in different
toplevel files, since the scopes of the two are not enclosing. But that should
not matter since for toplevel definitions we do have a defined receiver, namely
the synthetic enclosing object. Only for extension methods it happened that
that object was referred to via a This reference instead of a TermRef. This is
allowed in general but fell through a missing case in this instance.

Fixes #12897