Add verbose mode to WConf / nowarn to display matching message filters
In "verbose" mode warnings are printed with a list of matching message
filters that can be used in `-Wconf` and/or `@nowarn`. Example:

```
-- [E000] Syntax Warning: Test.scala:5:12 --------------------------------------
5 |  def bar = try 1
  |            ^^^^^
  |            A try without catch or finally is equivalent to putting
  |            its body in a block; no exceptions are handled.
Matching filters for @nowarn or -Wconf:
  - id=E0
  - name=EmptyCatchOrFinallyBlock

longer explanation available when compiling with `-explain`
```

Verbose mode can be enabled in `-Wconf`, for example for all warnings
using `-Wconf:any:v`.

Howeber, changing compiler flags is slow. To make the workflow more
convenient verbose mode can be enabled by annotating the surrounding
code with `@nowarn("verbose")` or `@nowarn("v")`:

```
  @nowarn("v")
  def bar = try 1
```