Deskolemise patterns to suppress exhaustivity warnings
Ideally I'd want to not be destructive with this widen, but it seems to
already being done on the other side: in `signature` where the
components of the unapply result type are obtained the call to:

    mt.instantiate(scrutineeTp :: Nil).finalResultType

removes the more precise `blub.type` that was present in the signature.

So now, by widening the pattern the spaces cancel each other out,
thus the match is deemed exhaustive.