Fix join of values
After #12711, it is no longer the case that `Warm < ThisRef`.
The reason is that the class parameters for `ThisRef` are always hot,
while it is not the case anymore after #12711.

Actually even before #12711, it's not the case because the outers of
warm objects may not be hot. The theory does not have inner classes,
it thus does not suffer from the problem.