Update invariantDisjoint check
The previous check was too conservative, is we have a proof that A and B are
disjoint, then A and B in Inv[A] and Inv[B] are clearly different types, and
Inv[A] and Inv[B] are disjoint. Also the isSameType() && fullyInstantiated
check is quite primitive, it can still be useful in some cases.