REPL: emit parse warnings by reusing the reporter
The original report requires the -deprecation and -source:future flags,
for which there is seems to be no support in repl.ScriptedTests.  So
switched to a non-deprecation syntax warning, that isn't guarded by
-source.

I might come back and add "pragma" support to the REPL scripts so I can
move the test out of pending.

Co-authored-by: Seth Tisue <seth@tisue.net>