Deprecate ad hoc group names with Regex and .r
The deprecation was first proposed in
https://github.com/scala/scala/pull/4990 but had been rejected
because Scala.js did not support inline group names. Now that
Scala.js 1.7.0 has been released with full inline group name
support, there is no reason to perpetuate this API.

Unfortunately, we cannot actually put `@deprecated` on the
constructor of Regex with group names, since there is no
alternative that does not take any group name. We *could* deprecate
it anyway, with the replacement being to use `.r`, but perhaps that
goes a bit too far.