Reimplement constraint merging for correctness
The previous implementation simply combined the content of both
constraints into one, but this is not enough since it meant that bounds
were not propagated and so transitivity was violated. For example, when
merging a constraint containing `?S <: ?T` and one containing `?T <:
?R`, the result did not verify `?S <:< ?R` (see the unit tests added in
ConstraintsTest).

The new implementation simply starts with one set of constraints and
then adds the constraints from the other set one by one using `<:<`
which takes care of propagating bounds. This is likely to be more
expensive than the previous implementation but it turns out that
`TyperState#mergeConstraintWith` is a rare operation (after the previous
commit it is only called 43 times when compiling scala3-compiler), so
the difference shouldn't be significant.

This also incidentally fixes #12730 because the previous logic for
checking if merging succeeded was flawed.