Restore unreachable warnings by converting literals
The fix implemented to deal with primitive types auto-converting was
suppressing a series of legitimate unreachable warnings.  So we removed
that fixed and tried to mirror the auto-conversion of the type, such
that the space intersection calculation ends up with the correct
results.

In doing so we triggered reachability warnings being emitted by
sealed trait companion object (or enum) `ordinal` synthetic methods, in
some weird test cases.  So we added an `@unchecked` annotation to
suppress those.

Co-authored-by: Seth Tisue <seth@tisue.net>