Fix printing of OrderingConstraint
When printing a constraint, we print types which might
refer to type variables defined in the constraint, but the type printer
will rely on ctx.typerState.constraint to determine if these type
variables are instantiated, and this might be a different constraint
than the one we're trying to print, leading to an incorrect output.

This commit fixes this by temporarily setting ctx.typerState.constraint
to the current constraint when printing it, this required moving the
printing logic from OrderingConstraint to PlainPrinter.

At the same time, we drop the distinction between `toText` and
`contentsToString` (the former wrapped the printed output in
"Constraint(...)" and the latter didn't, now we never do) because
preserving it would have been complicated and it didn't seem worth it.

Also fix Ordering#toString to correctly print the bounds (the logic was
there but it was dead code).