Tighten condition when to emit a "unary method cannot have parameters" warning
Also, use a more precise term in the error message. Unary method generally means
something entirely different!

Generally, was it really worth it to go into so much trouble for a warning
for an edge case? The previous fix overshot, excluding things that
are OK. I still believe it would have been better to do nothing aboit #9142.
The time we have already spent on this and the number of code lines we use to
address the issue is in no relation to the benefit of the fix.