Try to extract name when backtick is missing
Previously, completion prefix would show up as an empty string if we had an unclosed backtick. Now, we try to extract the prefix from the file contents in that case.

An alternative approach would be to tranfer somehow the name string  in `nme.Error`, but I didn't see anything like that done before, so not sure how to properly address that.

Fixes https://github.com/lampepfl/dotty/issues/12514