Parse empty Java compilation units as EmptyTree
Before this commit, an empty Java compilation unit (one without a
package declaration or any top level declarations) was parsed as
`PackageDef(Ident(<empty>),List())` which resulted in position not set errors.

With this commit, an empty Java compilation unit is parsed as `EmptyTree`.
This is consistent with the parsing of empty Scala compilation units.

Fixes #13310