ordinal: drop unchecked with sealed abstract skips
In the `ordinal` method, drop the use of `@unchecked` by skipping
children that are sealed traits or sealed abstract classes and don't
have any anonymous subclasses.

The `@unchecked` I added in ee5a82f65f42236d142f0e4b588e4522d03af616.
In a PR to specifically with reachability I started similar warnings in
`ordinal` (this was before adding `@unchecked` was merged) so I'm hoping
with this patch to avoid the false positives properly.