Fix crash on uninstantiated field access
The uninstantiated access in the test case happens because the closure
passed to the HKTypeLambda constructor to construct the result type by
`HKTypeLambda.newLikeThis` calls `subst` which eventually calls
`HKTypeLambda#equals`, I don't see a way to prevent that at the source,
so make equality checking robust to uninstantatied fields.

Fixes #13513.