Fix #13518: Scala.js: fix primitive + string when the string is a TermRef.
Sometimes, the argument string's `tpe` can be a `TermRef`, in which
case it is not directly a ref to `StringClass`. That was triggering
the assertion.

The assertions and the calls to `makePrimitiveBox` have actually
not been needed for a long time. The IR supports the arguments to a
`String_+` operator to be of any type. So we fix the issue by
removing all of that and directly emitting the `String_+`.