Fix missing SymbolOccurrence for aliased symbol
fix https://github.com/lampepfl/dotty/issues/13556

Previously, we include symbols iff it's name present in source by
`namePresentInSource`.
However, if the symbol is aliased by, for example, renamed imports,
there's difference beteen the name of the symbol and the name in source
even though they are the same symbol.

What we really need to exclude are the symbols that never appear in
source, and the gap between symbol name and name in source is
acceptable.

Therefore, this commit try to exlucde symbol using `isZeroExtent`
instead of `namePresentInSource`