Change syntax of erased function types and literals
It's now

    erased A => B
    erased (x: A) => B

instead of

    (erased A) => B
    (erased x: A) => B

The new syntax makes it clearer that the whole function is erased instead of just its parameters.
It's also a bit more regular. `erased` aligns with `implicit` for closures. The old syntax was the
last remnant where a keyword went inside the parens of a function type or literal.