Name mangle methods with erased context results
Add a `$` to the name of a method that has erased context results and that
may override some other method. This is to prevent the two methods from having
the same names and parameters in their erased signatures. We need a bridge between the
two methods, so they are not allowed to already override after erasure.