Remove unnecessary call to constrainResult
This was originally added in #1395 to make the following compile:

    (1, x => 2): (Int, Int => Int)

But reverting this change doesn't break anything anymore: the other call
to `constrainResult` in `Application#init` is enough to propagate the
constraint needed to type lambda. The strange thing is that this call
already existed before #1395, I don't know why it wasn't enough back then.