Don't follow BaseType of abstract binders in MT reduction
Fix #11982 and the associated soundness problem. The issue with the
behavior on master arises from the fact that type binder of match types
might change as context gets more precise, which results in a single
match type reducing in two different ways. This issue comes from the
fact that subtyping looks into base types, and is thus able to match a
type such as `T <: Tuple2[Int, Int]` against a pattern `case Tuple2[a,
b]`, even if the best solutions for `a` and `b` in the current context
are not guaranteed to be the best solution in more precise contexts
(such as at call site in the added test case).