Fixes and improvements to trace logging
The String.valueOf is because I somehow ended up with a null.

Then I found some docs in the history of trace, as well as a mistake in
a commit revert: using the non-StoreReporter logctx!

I also added a "log" variant to "force", which is also default enabled
but doesn't force the output to the standard out, meaning I can
bootstrap the compiler without it taking forever to output all the trace
output it generates while bootstrapping the compiler, bottlenecking on
console IO)...

Also, remove some dead code in Definitions.