move implementation of ordinal in enums to posttyper
also add a check that ordinal is not implemented
by the user, or mixed in by a trait - this is
necessary as scala.deriving.Mirror.Sum
delegates to ordinal method on an enum.
Note that before this commit, as enum cases would
previously declare ordinal methods at desugaring (without
an override flag), refchecks would issue an
override-without-override-modifier error anyway.

Co-authored-by: Jamie Thompson <bishbashboshjt@gmail.com>
Co-authored-by: Martin Odersky <odersky@gmail.com>