Use `inline val` for inlined `final val`s
These `val`s are already inlined and do not generate a field, but they
still generate a getter that is never used. My marking these as `inline`
we can remove the getter and save a few bytes at runtime.

By making these fields `inline` we can also check that the RHS constant
folds to a value at compile time.