Fix reachability by retaining spaces of Prod params
When intersecting a product type space with a type space, if the two
types aren't related, we must retain the spaces of the parameters,
otherwise the intersection will represent a bigger space of values that
it should (leading to a false positive unreachability warning).

I actually think that the first condition, `isSubType(tp1, tp2)`, should
be true, because `Bar` is a subtype of `Foo[X]`, provided `X` is
instantiated to `Nothing`.  Maybe that's not isSubType but something
else (because we want forSome X, not forAll X).  But, either way, fixing
that the intersection doesn't overpromise (i.e. lie) is a good fix
anyways.