GroupedIterator uses ArrayBuilder
Previous use of ArrayBuffer would shift the overlap from stepping,
fill the remainder, then toArray for result, then either retain
the buffer if stepping or clear it.

Now it allocates the result, copies the overlap, fills and returns it.
If stepping, an array of just overlap size is allocated and retained.
The retained buffer is never cleared, but reference is nulled when done.