Scala2Unpickler: Fix annotation argument handling
When an annotation argument is an Array, we can't know the element type
of the Array until we've applied the argument to the annotation
constructor, so we need to unpickle annotation arguments as untyped
trees and rely on resolveConstructor to type them. This is the same
logic we already use when unpickling Java classfiles since
610450f4f91be6c416a41ee2887f667e21385425.

Fixes #14223.