Dealias less in OrderingConstraint.replace and AvoidMap
Basically if the transformation doesn't apply to the dealias type,
then return the original type instead of the dealiased one.

In general it's reduce dealiasing to a minimum when typing,
because it can affect type unification as shown in the tests.

Unfortunately this doesn't handle the case of a singleton type alias,
because `widenSingletons` also dealiases and the same fix doesn't work.