Allow erased definitions without implementation
This is to avoid the need for a dummy implementation that is usually redundant and meaningless. To avoid this with a simple mechanism/rule that desugars abstract erased definitions into concrete ones.

The desugaring works for `val`/`def` as follows
```diff
- erased val f: T
+ erased val f: T = scala.compiletime.erasedValue
```
```diff
- erased def f: T
+ erased def f: T = scala.compiletime.erasedValue
```

This works well because of three properties of erased definitions
* An abstract erased definition is semantically equivalent to a non-abstract one. It is only their signature that influences their semantics.
* The implementation of an erased definition is only used to infer the type of the definition. Most of the time this is redundant.
* Most of the time the result type should be explicitly given