Properly desugar `inline given .. with ..`
```scala
inline given Foo with { .. }
```

now becomes
```scala
inline given def given_Foo: Foo with {..} = new Foo
class given_Foo extends Foo with {  }
```

There are no creation of `lazy vals` when we have `inline`.
We inline the `new givne_Foo` only, the class does not get inlined.
This is useful to remove the instatiation of the given instance when
the methods are inline themselfs.

Fixes #14282
Fixes #14177