Fix "have same type after erasure" check
Perform the overriding checks after elimByName. I observed some problem with
catsEffect2, where a super accessor method with a `() ?=> T` parameter was compared with a
corresponding super accessor method with a `=> T` parameter. One of these methods was generated
before elimByName, the other after. So comparing them at phase elimRepeated + 1 gave two different
types. The problem is fixed by comparing after elimByName, which means that the type of the second
method is converted to match the first.