community build: fix exclusion of flaky scalatest tests
The previous attempt at excluding the flaky tests, adapted from the
Scala 2 community build, was not fully working here, most likely
because of how those test sources are generated in the dotty version
of the ScalaTest build.

The fix in this commit has been verified by running the community build
with the sbt logging level set to info and inspecting the output.