Re-enable missing cmdTests
It might have been disabled by mistake because bootstrapCmdTests was
assumed to be the set of tests for the bootstrapped compiler.
The second is an extension for the second with tests that can only
be run on the bootstrapped compiler. It is now renamed to
bootstrappedOnlyCmdTests to make sure we avoid this potential confusion.

Fixes #14428