Search all source trees for a given span
When we create a synthetic companion object, for exmaple for enums, all the value definitions are put into it and the there are actually multiple trees that contain the position we are itnerested in.

Now, NavigateAST.pathTo takes in multiple trees since the method is already able to find the best fit.