Fix backend assertion when using an exported enum
An exported enum is a <static> final method of the enum object.  When
selecting an inner enum case from it, that method will be the prefix.
The method doesn't have a type symbol, we have to widen to method type
and reach for its result type before we can get a handle on the type
symbol to key off of.

I wonder whether this widening and going to the result type is something
that should happen during Erasure's denotation transformation though...

Co-authored-by: Seth Tisue <seth@tisue.net>