Allow applications in export qualifiers
This is a PR for a language extension to allow applications in export qualifiers.

## Motivation

The main motivation is to replace a common use case where today we still need an
implicit class: bulk decoration of method definitions from some template class.
To do this we'd need a wildcard export in an extension method clause. E.g.
```scala
class Ops(x: C) { ... }

extension (x: C)
  export Ops(x).*
  ...
```

The export clause here has a qualifier that is syntactically an application. Such
qualifiers are supported by this PR.