Fix "forward reference from self constructor" checking
The previous check did not consider the case where the self constructor
was itself a block that originated from named and default parameters in
the actual self constructor call. That gave a false negative for some code
in akka. The negative was not discovered before since we did not propagate
the correct context information into the expression of a block.