fix: ensure you still get completion when using backticks
This commit addresses the situation where you are trying to use a
backtick when it's not needed. For example if you have:

```scala
object Foo:
  case `bar`
  case foo

Foo.`b
```

You still expect to get a completion bar being backticked. This also
ensure that if you try ``Foo.`fo`` you also still get a completion.