Show case class signature when unapply is synthetic
Previously we would only show the unapply method's signature, which is not that useful for the user when writing pattern matches. Now, we should the case class signature in case of synthetic unapplies.

If this is the right direction, we could also solve this for actual unapplies and shwo the return value as signature isntead.