fix: don't split :<word> into two tokens.
In the situation where we have `:im<tab>` normally `:` and `im` would be split into
two tokens and it would result in `::imports`. Instead, we detect early if we are
about to complete a command and instead of parsing it normally like Scala we just
grab the entire thing as the word.