fix(completions): support backticked imports in completions
This will ensure that when you try to backtick an import whether it's
needed or not that you correctly capture the prefix and offer a
backticked completion. For example:

```scala
import scala.util.chaining.`s<TAB>
```

Results in an `Ident(<error>)` which we now check to see if a backtick
is causing it and correctly offer the completion.

Fixes #12514