Tolarate some faults when copying trees in InlineTreeMap
As a preparatory step the inliner maps the inlined body with a tree type map.
This map adjusts the tree and at the same time sets up the environment for
typing the tree. But this has the potential that re-typing during copying
will fail since the environment is not yet set up correctly. We avoid the
problem by ignoring a specific failure (function type in application does
not exist) and proceeding with the previous type.