Don't trust @unchecked types when optimizing type tests
Some type tests are elided or replaced with non null-checks if we can show
that the expression's type is a subtype of the test type. But that can be done
safely only if the expression's type does not come from an `@unchecked` cast,
inserted by pattern matching or the user.

Fixes #14705