Fix hoisting local symbols
The previous algorithm computed a "logical owner" of term-owned symbols
from dependencies. But it would switch back from the computed logical
owner to the enclosing class under certain conditions. The problem is
that such a decision cannot be done independently for each symbol, since
it also affects dependent symbols. We now move this distinction into
the logical owner initialization itself, instead of doing it as a
post processing step.

Fixes #14707