Improve overloading resolution if expected type is not FunProto
If expected type is a selection proto or similar we picked all alternatives
that were compatible with it. But if the expected type was meant to be
used after extension method expansion, then no alternative would be classified
as compatible.

We now drop all method alternatives in this case. If only one alternative remains,
we pick that one.

Fixes #14279