refactor: change Message kind into an enum instead of a String
While digging into Diagnostics there wasn't an easy way to tell all the
available `kind`s that there could be since they were just a string
value. So this small changes just makes the `kind` field in `Message` a
`MessageKind` which is an enum of all the types of message kinds that
exist.