feat: add -h to alternatives for -help
I've tried this no less than 3 times over the weekend and it annoys me
every time I get the `bad option '-h' was ignored` and it brings me
into the REPL. I think it's always a good idea to just default
`-help|--help|-h` all to help. Since this doesn't seem to conflict with
anything this just adds it in.