Indy-fy case class `hashCode`
  1. `SyntheticMembers` uses `CaseClassMethods.caseHashCode(this)` as
     the implementation of `hashCode`

  2. JVM backend recognizes `CaseClassMethods.caseHashCode(this)` and
     replaces it with `InvokeDynamic #0:hashCode:(LMyCaseClass;)I` with
     bootstrap method `CaseClassMethods` and bootstrap arguments
     including the case accessors

  3. `CaseClassMethods.bootstrap` generates a `MethodHandle` with the
     same semantics as the manually generated bytecode

This isn't yet expected to be robust, but it does return the right
hashCode for toy case class examples. The idea follows @sjrd's
suggestion of using a "magic" method. In order to make the typechecker
happy, there actually _is_ a valid implementation of this method too.

I'm still unsure how to proceed around the JS backend