fix: ensure DeprecatedWithOperator is shown to user when used.
Right now a user can use a `with` operator without any deprecation
notice when this has been deprecated for quite some time. I think this
was missed recently with some of the changes to sourceVersion and this
was still checking against `future`. I believe this should just be
removed and always shown now correct?