Fix constToLiteral
We previously converted an expression with constant type to a literal if
the expression was idempotent. This can hide side effects in the case where
the expression is a selection from an object or lazy val. Demanding purity
instead prodcues tons of errors involving inline vals on objects.

We now demand idempotency if the expression refers to an inline val (or
an operation over an inline val), and purity elsewhere.

Fixes #2266