Fix #14773: Reuse the param slots for the tailrec local mutable vars.
The tailrec phase generates code that looks like

  class Bar {
    def foo(x1: Int, x2: Int): Int = {
      var this$: Bar = this;
      var taillocal$x1: Int = x1;
      var taillocal$x2: Int = x2;
      ...
      // body where `this`, `x1` and `x2` are replaced by
      // `this$`, `taillocal$x1` and `taillocal$x2`
    }
  }

This generates bytecode where the `this` value and the parameters
never actually change, and are never used. Instead, the synthetic
mutable variables are used instead.

As described in the linked issue, this confuses debuggers, which
only display the never-changing original `this` value and
parameters.

In this commit, we intercept this shape of code in the back-end.
We reliable identify tailrec-generated `ValDef`s from their
semantic names, with an additional safety check that they are
`Synthetic | Mutable`. When we find this shape, we do not allocate
local slots for the `var`s, and instead reuse the slots for `this`
and the parameters. We skip past the `ValDef`s so that code
generation does not re-emit useless assignments.