Supress safe-init warnings for NameKind HashMaps
This adds the `@unchecked` annotation to each of the four lines that add the object under construction to the relevant HashMap.

In the `NameKinds.scala` file, all of the `NameKind`s are created and added to one of four HashMaps based on their superclass. The lines that add the NameKind to the HashMaps are at the end of the relevant superclasses. This leads to a safe-init error since subclasses can add further fields and initialization logic that gets run after these lines.

These warnings could be eliminated without the use of `@unchecked`, but it would require redundantly adding the same line to all `NameKind` subclasses that get initialized. This means that anyone who changes this code in the future would need to remember to do this for all of the relevant subclasses. This could also be fixed without `@unchecked` if there were a mechanism like `DelayedInit` in Scala3.

Thus, instead of making a more involved change that would render the code redundant, overly verbose, and less maintainable, I opted for using `@unchecked` instead.