Disallow old uses of quoted and spliced types
These have have been emitting migration warning in 3.0 and 3.1 even if the
syntax was already deprecated in 3.0. These where allowed syntax found
in older papers.

The case `'[T]` was actually buggy but no one reported this bug so far.
A good indication that this older syntax has not been used in a while.

First step towards fixing #15009.